+++ lib/refresh/refresh_config.dart
@@ -2,6 +2,7 @@ import 'package:flutter/widgets.dart';
 
 import 'refresh_controller.dart';
 
+/// 刷新状态
 enum RefreshStatus {
   idle,
   prepared,
@@ -9,9 +10,19 @@ enum RefreshStatus {
   completed,
 }
 
-typedef RefreshIndicatorBuilder = Widget Function(BuildContext context, RefreshStatus status);
-typedef RefreshStatusDidChange = void Function(RefreshController controller, RefreshStatus status);
+/// 刷新组件创建
+typedef RefreshIndicatorBuilder = Widget Function(
+  BuildContext context,
+  RefreshStatus status,
+);
 
+/// 刷新状态回调
+typedef RefreshStatusDidChange = void Function(
+  RefreshController controller,
+  RefreshStatus status,
+);
+
+/// tableView刷新控制类
 class RefreshConfig {
   RefreshIndicatorConfig headerConfig;
   RefreshIndicatorConfig footerConfig;
@@ -36,11 +47,15 @@ class RefreshConfig {
   }) : controller = controller ?? RefreshController();
 }
 
+/// 刷新组件配置
 class RefreshIndicatorConfig {
+  /// 触发刷新距离
   final double triggerDistance;
 
+  /// 完成刷新后，停留在completed状态的停留时长
   final int completeDuration; // in millisecond
 
+  /// 可见范围
   final double visibleRange;
 
   const RefreshIndicatorConfig({