+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 ## 1.0.0
 
 * Create CupertinoTableView with Refresh Header&Footer
+
+
+## 1.0.1
+
+* 代码优化
+1. 调整设置pressedOpacity的方法，现在放到delegate中设置
+2. 去掉多余的依赖库
+3. 以来dart版本降级，现在最低支持dart2.12
+4. 补充部分注释和readme
+5. 修复当不设置section decoration时的异常